chore(theme context): shifted the useResumeTheme hook to context file.
