+++ src/components/resume-builder-v2/context/ResumeThemeContext.tsx
@@ -1,5 +1,5 @@
 'use client';
-import { createContext, useState } from 'react';
+import { createContext, useContext, useState } from 'react';
 
 // TYPES
 import type { ThemeColorKeys, ThemeColorValues } from '../types/theme';
@@ -33,4 +33,14 @@ export const ResumeThemeProvider: React.FC<{ children: React.ReactNode }> = ({
       {children}
     </ResumeThemeContext.Provider>
   );
+};
+
+export const useResumeTheme = () => {
+  const context = useContext(ResumeThemeContext);
+
+  if (!context) {
+    throw new Error('useResumeTheme must be used within a ThemeProvider');
+  }
+
+  return context;
 };
